Create phonebook.c
The=is program works as a digital phone book, in which the user can store the names and numbers of  their contacts. A CSV file gets updated in real time that can be also used in programs like Microsoft Excel. Also the CSV file gets created after compiling the project too and the printed contacts are accessible directly via the program too.